+++ front-end/src/components/Header/index.js
@@ -4,13 +4,14 @@ import SideBar from '../SideBar/index';
 
 import './style.css';
 
-const Header = ({ title }) => {
+const Header = ({ title, userType }) => {
+
   return (
     <div>
       <header className="title">
         <h1 data-testid="top-title">{title}</h1>
       </header>
-      <SideBar />
+      <SideBar role={userType} />
     </div>
   );
 };
