+++ front-end/src/components/Login.js
@@ -2,6 +2,7 @@ import React, { useState } from 'react';
 import { Link, useHistory } from 'react-router-dom';
 // import { TrybeerContext } from '../context';
 import API from '../services/api';
+import Header from './Header';
 
 const Login = () => {
   // const { user, setUser } = useContext(TrybeerContext);
@@ -45,53 +46,56 @@ const Login = () => {
   };
 
   return (
-    <form onSubmit={handleSubmit} className="m-3">
-      <section className="card w-75 mx-auto m-3">
-        <div className="form-group w-75 mx-auto m-2">
-          <label htmlFor="email">Email</label>
-          <input
-            onChange={(e) => setEmail(e.target.value)}
-            data-testid="email-input"
-            type="text"
-            name="email"
-            id="email"
-            required
-            className="form-control"
-          />
-        </div>
+    <div>
+      <Header title={'Trybeer'} />
+      <form onSubmit={handleSubmit} className="m-3">
+        <section className="card w-75 mx-auto m-3">
+          <div className="form-group w-75 mx-auto m-2">
+            <label htmlFor="email">Email</label>
+            <input
+              onChange={(e) => setEmail(e.target.value)}
+              data-testid="email-input"
+              type="text"
+              name="email"
+              id="email"
+              required
+              className="form-control"
+            />
+          </div>
 
-        <div className="form-group w-75 mx-auto m-2">
-          <label htmlFor="password">Password</label>
-          <input
-            onChange={(e) => setPassword(e.target.value)}
-            data-testid="password-input"
-            type="password"
-            name="password"
-            id="password"
-            minLength="6"
-            required
-            className="form-control"
-          />
-        </div>
+          <div className="form-group w-75 mx-auto m-2">
+            <label htmlFor="password">Password</label>
+            <input
+              onChange={(e) => setPassword(e.target.value)}
+              data-testid="password-input"
+              type="password"
+              name="password"
+              id="password"
+              minLength="6"
+              required
+              className="form-control"
+            />
+          </div>
 
-        <span className="mx-auto m-3 text-danger">{errorMsg.toUpperCase()}</span>
+          <span className="mx-auto m-3 text-danger">{errorMsg.toUpperCase()}</span>
 
-        <div className="mx-auto m-2">
-          <button
-            data-testid="signin-btn"
-            type="submit"
-            disabled={disableButton}
-            className="btn btn-warning m-2"
-          >
-            ENTRAR
-          </button>
+          <div className="mx-auto m-2">
+            <button
+              data-testid="signin-btn"
+              type="submit"
+              disabled={disableButton}
+              className="btn btn-warning m-2"
+            >
+              ENTRAR
+            </button>
 
-          <Link data-testid="no-account-btn" to="/register" className="m-2">
-            Ainda não tenho conta
-          </Link>
-        </div>
-      </section>
-    </form>
+            <Link data-testid="no-account-btn" to="/register" className="m-2">
+              Ainda não tenho conta
+            </Link>
+          </div>
+        </section>
+      </form>
+    </div>
   );
 };
