+++ front-end/src/components/UserProfile/style.css
@@ -1,8 +1,4 @@
 form {
   padding: 25%;
   padding-top: 10%;
-<<<<<<< HEAD
 }
-=======
-}
->>>>>>> 67290b957dc196198b362b2ea3a09d428b7ce878
