+++ front-end/src/components/UserRegister.js
@@ -1,6 +1,8 @@
 import React, { useState } from 'react';
 import { useHistory } from 'react-router-dom';
 
+import Header from './Header';
+
 import API from '../services/api';
 
 const UserRegister = () => {
@@ -49,71 +51,74 @@ const UserRegister = () => {
   };
 
   return (
-    <form onSubmit={handleSubmit}>
-      <section className="card w-75 mx-auto m-3">
-        <div className="form-group w-75 mx-auto m-2">
-          <label htmlFor="name">Nome</label>
-          <input
-            data-testid="signup-name"
-            id="name"
-            name="name"
-            type="text"
-            onChange={(e) => setName(e.target.value)}
-            minLength="12"
-            className="form-control"
-          />
-        </div>
+    <div>
+      <Header title={'Register'} usertype={'client'}/>
+      <form onSubmit={handleSubmit}>
+        <section className="card w-75 mx-auto m-3">
+          <div className="form-group w-75 mx-auto m-2">
+            <label htmlFor="name">Nome</label>
+            <input
+              data-testid="signup-name"
+              id="name"
+              name="name"
+              type="text"
+              onChange={(e) => setName(e.target.value)}
+              minLength="12"
+              className="form-control"
+            />
+          </div>
 
-        <div className="form-group w-75 mx-auto m-2">
-          <label htmlFor="email">Email</label>
-          <input
-            data-testid="signup-email"
-            id="email"
-            name="email"
-            type="email"
-            onChange={(e) => setEmail(e.target.value)}
-            className="form-control"
-          />
-        </div>
+          <div className="form-group w-75 mx-auto m-2">
+            <label htmlFor="email">Email</label>
+            <input
+              data-testid="signup-email"
+              id="email"
+              name="email"
+              type="email"
+              onChange={(e) => setEmail(e.target.value)}
+              className="form-control"
+            />
+          </div>
 
-        <span className="mx-auto m-3 text-danger">{errMsg}</span>
+          <span className="mx-auto m-3 text-danger">{errMsg}</span>
 
-        <div className="form-group w-75 mx-auto m-2">
-          <label htmlFor="password">Password</label>
-          <input
-            data-testid="signup-password"
-            id="password"
-            name="password"
-            type="password"
-            onChange={(e) => setPassword(e.target.value)}
-            minLength="6"
-            className="form-control"
-          />
-        </div>
+          <div className="form-group w-75 mx-auto m-2">
+            <label htmlFor="password">Password</label>
+            <input
+              data-testid="signup-password"
+              id="password"
+              name="password"
+              type="password"
+              onChange={(e) => setPassword(e.target.value)}
+              minLength="6"
+              className="form-control"
+            />
+          </div>
 
-        <div className="form-group w-75 mx-auto m-2">
-          <input
-            data-testid="signup-seller"
-            id="sell"
-            name="sell"
-            type="checkbox"
-            onChange={(e) => setCheckbox(e.target.value)}
-          />
-          <label htmlFor="sell" className="m-2">
-            Quero Vender
-          </label>
-        </div>
+          <div className="form-group w-75 mx-auto m-2">
+            <input
+              data-testid="signup-seller"
+              id="sell"
+              name="sell"
+              type="checkbox"
+              onChange={(e) => setCheckbox(e.target.value)}
+            />
+            <label htmlFor="sell" className="m-2">
+              Quero Vender
+            </label>
+          </div>
 
-        <button
-          data-testid="signup-btn"
-          type="submit"
-          disabled={disableButton}
-          className="btn btn-warning w-50 mx-auto m-2"
-        >
-          Cadastrar
-        </button>
-      </section>
-    </form>
+          <button
+            data-testid="signup-btn"
+            type="submit"
+            disabled={disableButton}
+            className="btn btn-warning w-50 mx-auto m-2"
+          >
+            Cadastrar
+          </button>
+        </section>
+      </form>
+    </div>
   );
 };
